09/18 DB 연동 테스트

+++ src/main/java/kr/ac/kopo/viewfit/entity/Manufacturer.java
@@ -12,12 +12,12 @@ public class Manufacturer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column
     private String name;
 
     @Column
-    private int type;
+    private Integer type;
 
 }
